upd structure
