+++ hypex/analyzers/base.py
