+++ hypex/experiments/ab.py
@@ -1,6 +1,6 @@
 from hypex.analyzers.ab import ABAnalyzer
-from hypex.comparators.comparators import GroupDifference, GroupSizes, GroupATE
-from hypex.comparators.hypothesis_testing import TTest, MannWhitney
+from hypex.comparators.comparators import GroupDifference, GroupSizes, ATE
+from hypex.comparators.hypothesis_testing import TTest, UTest
 from hypex.dataset.roles import TargetRole, TreatmentRole
 from hypex.experiments.base import Experiment, OnRoleExperiment
 
@@ -10,11 +10,11 @@ AB_TEST = Experiment(
         OnRoleExperiment(
             executors=[
                 GroupDifference(grouping_role=TreatmentRole()),
-                GroupATE(
+                ATE(
                     grouping_role=TreatmentRole(),
                 ),
                 TTest(grouping_role=TreatmentRole()),
-                MannWhitney(grouping_role=TreatmentRole()),
+                UTest(grouping_role=TreatmentRole()),
             ],
             role=TargetRole(),
         ),
