+++ hypex/reporters/aa.py
@@ -1,12 +1,12 @@
 from typing import Dict, Any
 
+from hypex.analyzers.aa import OneAASplitAnalyzer
+from hypex.comparators.comparators import GroupDifference, GroupSizes
 from hypex.dataset.dataset import ExperimentData
-from hypex.reporters.reporters import DictReporter
+from hypex.reporters.base import DictReporter
 from hypex.splitters.aa import AASplitter
 from hypex.utils.constants import ID_SPLIT_SYMBOL
 from hypex.utils.enums import ExperimentDataEnum
-from hypex.comparators.comparators import GroupDifference, GroupSizes
-from hypex.analyzers.aa import OneAASplitAnalyzer
 
 
 class AADictReporter(DictReporter):
@@ -18,7 +18,9 @@ class AADictReporter(DictReporter):
         return int(aa_id) if aa_id.isdigit() else None
 
     def extract_group_difference(self, data: ExperimentData) -> Dict[str, Any]:
-        group_difference_ids = data.get_ids(GroupDifference)[GroupDifference][ExperimentDataEnum.analysis_tables.value]
+        group_difference_ids = data.get_ids(GroupDifference)[GroupDifference][
+            ExperimentDataEnum.analysis_tables.value
+        ]
         t_data = data.analysis_tables[group_difference_ids[0]]
         for aid in group_difference_ids[1:]:
             t_data = t_data.append(data.analysis_tables[aid])
